añadir, modificar y borrar mis clases, mis ranking y clases del admin

+++ src/main/java/com/example/gymweb/RestControllers/LessonsRest.java
@@ -4,8 +4,10 @@ import com.example.gymweb.Services.LessonsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import java.util.Collection;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api")
@@ -14,8 +16,8 @@ public class LessonsRest {
     @Autowired
     LessonsService lessonsService;
     //to get lesson by id
-    @GetMapping("/{id}")
-    public ResponseEntity<Lesson> getLesson(@PathVariable long id) {
+    @GetMapping("/lesson/{id}")
+    public ResponseEntity<Lesson> getLessonById(@PathVariable long id) {
         Lesson lesson = lessonsService.getLessonById(id);
         if (lesson != null) {
             return new ResponseEntity<>(lesson, HttpStatus.OK);
@@ -24,22 +26,35 @@ public class LessonsRest {
         }
     }
     //to get all lessons
-    @GetMapping("/")
+    @GetMapping("/lesson")
     public ResponseEntity<Collection<Lesson>> getAllLessons() {
         Collection<Lesson> lessons = lessonsService.getLessons();
         return new ResponseEntity<>(lessons, HttpStatus.OK);
     }
     //to create a lesson
-    @PostMapping("/")
+    @PostMapping("/lesson")
     public ResponseEntity<Lesson> createLesson(@RequestBody Lesson lesson) {
-        Lesson createdLesson = lessonsService.addLesson(lesson);
-        return new ResponseEntity<>(createdLesson, HttpStatus.CREATED);
+        lessonsService.addLesson(lesson);
+        return new ResponseEntity<>(lesson, HttpStatus.CREATED);
     }
-    //this code is for to delete a lesson
-    @DeleteMapping("/{id}")
+    //this code is for deleting a lesson
+    @DeleteMapping("/lesson/{id}")
     public ResponseEntity<?> deleteLesson(@PathVariable long id) {
+        Lesson lesson = lessonsService.getLessonById(id);
+        if (lesson == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         lessonsService.deleteLessonById(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+    //update an existing lesson giving his ID
+    @PutMapping("/lesson/{id}")
+    public ResponseEntity<Lesson> updateLesson(@RequestBody Lesson lesson,@PathVariable long id, BindingResult bindingResult){
+        if (bindingResult.hasErrors()){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        lessonsService.updateBook(id,lesson);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
 }
